% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificial_risk.R
\name{artificial_risk}
\alias{artificial_risk}
\title{Simulate spatially dependent random data}
\usage{
artificial_risk(shapefile, n_convolution, rate_per_100k, seed_risk,
  seed_poisson)
}
\arguments{
\item{shapefile}{a SpatialPolygonsDataFrame with data columns region_ID (= AGS-Key) and n (= population)}

\item{n_convolution}{number, how often local effect should be leveled with its neighbours}

\item{rate_per_100k}{average risk per 100k people}

\item{seed_risk}{seed for simulating theta}

\item{seed_poisson}{seed for simulatijng case numbers}
}
\value{
a SpatialPolygonsDataFrame with additional data columns u, v, mu, theta, O, E
}
\description{
Simulate spatially dependent random data
}
\details{
The BYM-model assumes data to be like O_i ~ Poi(E_i * exp(u_i + v_i + mu))
    with observed (O) and expected (E) cases and random effects mu (constant), v (unique) and u (local).
    Exactly this model is simulated.
    The autocorrelation of u is simulated via averaging over its neighbours n_convolution times
}
