% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bugs.R
\name{model_incidence}
\alias{model_incidence}
\title{OpenBUGS-model for BYM model}
\usage{
model_incidence()
}
\value{
the model description in a character element (multiple lines separated by "\ n")
}
\description{
OpenBUGS-model for BYM model
}
