% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stan.R
\name{model_STAN_bym2}
\alias{model_STAN_bym2}
\title{Rstan-model for BYM-2 model}
\usage{
model_STAN_bym2()
}
\value{
the model description in a character element (multiple lines separated by "\ n")
}
\description{
Rstan-model for BYM-2 model
}
