% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stan.R
\name{myrstan_indexByComponent}
\alias{myrstan_indexByComponent}
\title{myrstan_indexByComponent}
\usage{
myrstan_indexByComponent(x)
}
\arguments{
\item{x}{vector of component ids}
}
\value{
vector of per-component consecutive node ids
}
\description{
myrstan_indexByComponent
}
\details{
subfunction of myrstan_scale_nb_components
}
