% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stan.R
\name{myrstan_nb2graph}
\alias{myrstan_nb2graph}
\title{myrstan_nb2graph}
\usage{
myrstan_nb2graph(x)
}
\arguments{
\item{x}{nb_object}
}
\value{
data.frame containing num nodes, num edges,
}
\description{
myrstan_nb2graph
}
