% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stan.R
\name{myrstan_nb2subgraph}
\alias{myrstan_nb2subgraph}
\title{myrstan_nb2subgraph for a given subcomponent, return graph as lists of node1, node2 pairs}
\usage{
myrstan_nb2subgraph(x, c_id, comp_ids, offsets)
}
\arguments{
\item{x}{nb object}

\item{c_id}{subcomponent id}

\item{comp_ids}{vector of subcomponent ids}

\item{offsets}{vector of subcomponent node numberings}
}
\value{
list of node1, node2 ids
}
\description{
myrstan_nb2subgraph for a given subcomponent, return graph as lists of node1, node2 pairs
}
\details{
subfunction of myrstan_scale_nb_components
}
