% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stan.R
\name{myrstan_scale_nb_components}
\alias{myrstan_scale_nb_components}
\title{myrstan_scale_nb_components}
\usage{
myrstan_scale_nb_components(x)
}
\arguments{
\item{x}{nb_object}
}
\value{
vector of per-component scaling factor (for BYM2 model) scaling factor for singletons is 0
}
\description{
myrstan_scale_nb_components
}
