% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plotting and colouring a map}
\usage{
plot_map(formula, shapefile, title = "")
}
\arguments{
\item{formula}{a formula in the form plot_value ~ region_ID. Values of plot_value less (or equal to) 0 are ignored in the calculation of colorings and are assigned lowest coloring.}

\item{shapefile}{the SpatialPolygonsDataFrame shapefile for plotting with color values in data slot}

\item{title}{caption of legend as character string}
}
\value{
a tmap object with the coloured map
}
\description{
Plotting and colouring a map
}
\details{
creates a map with the information passed in the data slot.
    the coloring scheme is viridisLite::cividis.
    Coloring steps are geometrically distributed.
    if the (non-zero) minimum is less tha 1, the colors are centered at 1
}
