% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape_GER}
\alias{shape_GER}
\title{Geographical data for Germany}
\format{A SpatialPolygonsDataFrame with 402 polygons and 2 variables:
\describe{
  \item{region_ID}{Amtlicher Gemeindeschlüssel, see https://en.wikipedia.org/wiki/Community_Identification_Number#Germany}
  \item{n}{average population in the time interval from 2008 to 2015}
}}
\usage{
shape_GER
}
\description{
Geographical data for Germany
}
\details{
Access via load(file.path(find.package("compareBYMimplementations"), "data", "shape_GER.Rdata"))
}
\keyword{datasets}
