% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_bugs.R
\name{smooth_bugs}
\alias{smooth_bugs}
\title{Smoothing map with MCMC (OpenBUGS) and BYM model}
\usage{
smooth_bugs(formula, data, shapefile, BUGS_tempdir, time_scaling = 1,
  ...)
}
\arguments{
\item{formula}{a formula in one of the form
\itemize{
 \item{}{observed ~ expected + region_ID}
}}

\item{data}{data.frame corresponding to formula}

\item{shapefile}{of class SpatialPolygonsDataFrame (see package sp),
where rownames of shapefile\@data equal region_ID in formula}

\item{BUGS_tempdir}{Some (empty) directory for OpenBUGS to store model information}

\item{time_scaling}{Openbugs uses 10,000 * time_scaling iterations,
First half is burnin}

\item{...}{more parameters of other methods (to unify call)}
}
\value{
A two element list
\itemize{
 \item{"SIR"}{A tibble (=data.fame) with region_ID and SIR (observed / expected)}
 \item{"run"}{the full OpenBUGS object for debugging}
}
}
\description{
Smoothing map with MCMC (OpenBUGS) and BYM model
}
