% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_inla.R
\name{smooth_inla}
\alias{smooth_inla}
\title{Smoothing map with INLA and BYM model}
\usage{
smooth_inla(formula, data, shapefile, INLA_model, ...)
}
\arguments{
\item{formula}{a formula in one of the form
\itemize{
 \item{}{observed ~ expected + region_ID}
}}

\item{data}{data.frame corresponding to formula}

\item{shapefile}{of class SpatialPolygonsDataFrame (see package sp),}

\item{INLA_model}{one of "bym" and "bym2" - defines, how INLA should parametrize the model}

\item{...}{more parameters of other methods (to unify call)
where rownames of shapefile\@data equal region_ID in formula}
}
\value{
A two element list
\itemize{
 \item{"SIR"}{A tibble (=data.fame) with region_ID and SIR (observed / expected)}
 \item{"run"}{the full INLA object for debugging}
}
}
\description{
Smoothing map with INLA and BYM model
}
