% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_rpkg.R
\name{smooth_rpkg}
\alias{smooth_rpkg}
\title{Smoothing map with R package CARBayrs and BYM model}
\usage{
smooth_rpkg(formula, data, shapefile, ...)
}
\arguments{
\item{formula}{a formula in one of the form
\itemize{
 \item{}{observed ~ expected + region_ID}
}}

\item{data}{data.frame corresponding to formula}

\item{shapefile}{of class SpatialPolygonsDataFrame (see package sp),
where rownames of shapefile\@data equal region_ID in formula}

\item{...}{more parameters of other methods (to unify call)}
}
\value{
A two element list
\itemize{
 \item{"SIR"}{A tibble (=data.fame) with region_ID and SIR (observed / expected)}
 \item{"run"}{the full CARBayes object for debugging}
}
}
\description{
Smoothing map with R package CARBayrs and BYM model
}
\details{
For number of MCMC-iterations, the default is choosen as in the CARBayes-Paper (2013)
}
