% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stan.R
\name{smooth_stan}
\alias{smooth_stan}
\title{Smoothing map with Rstan and BYM model}
\usage{
smooth_stan(formula, data, shapefile, STAN_dir, STAN_cores = NULL,
  STAN_bym_model, ...)
}
\arguments{
\item{formula}{a formula in one of the form
\itemize{
 \item{}{observed ~ expected + region_ID}
}}

\item{data}{data.frame corresponding to formula}

\item{shapefile}{of class SpatialPolygonsDataFrame (see package sp),}

\item{STAN_dir}{Some (empty) directory for Rstan to store model information}

\item{STAN_cores}{Number of cores for Rstan for parallel computing}

\item{STAN_bym_model}{precompiled stan bym2-model}

\item{...}{more parameters of other methods (to unify call)}
}
\value{
A two element list
\itemize{
 \item{"SIR"}{A tibble (=data.fame) with region_ID and SIR (observed / expected)}
 \item{"run"}{the full Rstan object for debugging}
}
}
\description{
Smoothing map with Rstan and BYM model
}
