% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_BYMresult.R
\name{summarize_BYMresult}
\alias{summarize_BYMresult}
\title{Calculate Coverage from "true" Simulation and BYM smoothing}
\usage{
summarize_BYMresult(estimation_tbl, simulation_tbl, method_name = NA,
  region = NA, n = NA, ID = NA, seed_risk = NA,
  seed_poisson = NA, n_convolution = NA, rate_per_100k = NA,
  duration = NA)
}
\arguments{
\item{estimation_tbl}{What should be estimated}

\item{simulation_tbl}{What was estimated}

\item{method_name}{the software used}

\item{region}{an identifier for the simulated region}

\item{n}{the number of counties in region}

\item{ID}{an ID of the try}

\item{seed_risk}{seed for risk distribution}

\item{seed_poisson}{seed for case distribution}

\item{n_convolution}{number of related neighbours}

\item{rate_per_100k}{determinant for case numbers per county}

\item{duration}{length of calculation time}
}
\value{
a data.frame with one row and many information
}
\description{
Calculate Coverage from "true" Simulation and BYM smoothing
}
