% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_population.R
\name{wrangle_population}
\alias{wrangle_population}
\title{Bereite Populationsdaten vor}
\usage{
wrangle_population(file, years = 1900:2050, verbose = TRUE)
}
\arguments{
\item{file}{Dateipfad zu Datei mit Bevölkerungszahlen}

\item{years}{Bilde Mittelwert über die hier genannten Jahre}

\item{verbose}{Soll hin und wieder ein Status auf der Konsole ausgegeben werden?}
}
\value{
ein data-frame mit Bevölkerungszahlen nach Kreis, Geschlecht und Altersgruppe
}
\description{
Bereite Populationsdaten vor
}
