% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_shapefile.R
\name{wrangle_shapefile}
\alias{wrangle_shapefile}
\title{Prepare GIS-Data}
\usage{
wrangle_shapefile(dir, layer, verbose = TRUE)
}
\arguments{
\item{dir}{Directory of Shapefiles}

\item{layer}{Name of Layer}

\item{verbose}{Report status on console?}
}
\value{
a SpatialPolygonsDataFrame
}
\description{
Prepare GIS-Data
}
